fix(sdk): use full import paths
