+++ packages/sdk/ts/proof/types.ts
@@ -1,9 +1,8 @@
+import type { ITallyData } from "../tally/types";
 import type { Groth16Proof, SnarkProof } from "@maci-protocol/contracts";
 import type { CircuitInputs } from "@maci-protocol/core";
 import type { BigNumberish, Signer } from "ethers";
 import type { PublicSignals } from "snarkjs";
-
-import { ITallyData } from "../tally";
 
 /**
  * Interface for the arguments to the proveOnChain command
